We're getting closer to a quasi-usable emgen
There are still room for improvement, and of course there will be bugs,
but step-by-step, it's getting there.

It also adds automatic release to github. Since there are no quality
gates (yet) it's a bit scary, but slowly there'll be improvements in the
development flow that will make it hard to screw up. Until then, just be
cautious.

- Fixes #7
- Add automatic release to .travis.yml